+++ release_notes.md
@@ -3,6 +3,7 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR)
 -->
+
 - Update Java Worker Version to [2.18.1](https://github.com/Azure/azure-functions-java-worker/releases/tag/2.18.1)
 - Introduced support for response compression, which can be enabled through explicit opt-in (#10870)
 - Add support for new FeatureFlag `EnableAzureMonitorTimeIsoFormat` to enable iso time format for azmon logs for Linux Dedicated/EP Skus. (#10684)
@@ -15,3 +16,4 @@
 - Increased the GC allocation budget value to improve cold start (#10953)
 - Fixed bug that could result in "Binding names must be unique" error (#10938)
 - Memory allocation optimizations in `RpcWorkerConfigFactory.AddProviders` (#10959)
+- Fix race condition that leads the host to initialize placeholder (warmup) function in Linux environments (#10848)